requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
Werkzeug==2.3.7
psycopg2-binary==2.9.8
python-dotenv==1.0.0
bcrypt==4.1.2
qrcode[pil]==7.4.2
python-escpos==3.0a9
gunicorn==21.2.0
openpyxl==3.1.2
reportlab==4.0.7